+++ lib/flutter_tyre.dart
@@ -3,15 +3,11 @@ library flutter_tyre;   //  ตั้งค่าให้เป็น Library  
 //  Import   Library  ที่ต้องการใช้
 
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+// import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 
 
 //   ใช้  part    เพื่ออ้างอิงที่อยู่ของไฟล์ ที่ต้องการทั้งหมด
 
-part './subwidgets/inkwell_menu.dart';
-part './subwidgets/search.dart';
-part './mainwidgets/top_header_menu.dart';
-part './widgets/top_left_header_menu.dart';
-part './widgets/top_right_header_menu.dart';
+part 'subwidgets/inkwell_text.dart';